Fix: Resolve widespread compiler warnings from -Wextra flag
This commit continues the systematic effort to enable clean compilation
with '-Wextra' and '-Werror'. A large number of files were updated to
resolve three main categories of warnings, improving code safety,
correctness, and portability.

- Resolved numerous signed vs. unsigned comparison warnings
  (`-Wsign-compare`), particularly in `for` loops and conditional
  logic involving `size_t`, `uvast`, and `pid_t` types. Fixes follow a
  standard pattern of checking for negative values before casting and
  comparing.

- Corrected intentional `switch` statement fall-throughs by replacing
  legacy comments with the compiler-recognized `/* FALLTHROUGH */`
  annotation to resolve `-Wimplicit-fallthrough` warnings.

- Fixed multiple instances of outdated C89-style struct initializers,
  using modern C99 designated initializers (`.field = value`) to prevent
  `-Wmissing-field-initializers` errors and make the code more robust.

+++ ici/libbloom-master/bloom.c
@@ -39,7 +39,8 @@ static int bloom_check_add(struct bloom * bloom,
   register unsigned int mask;
   register unsigned char c;
 
-  for (i = 0; i < bloom->hashes; i++) {
+    /* Cast bloom->hashes to unsigned to fix signedness warning. */
+    for (i = 0; i < (unsigned int)bloom->hashes; i++) {
     x = (a + i*b) % bloom->bits;
     byte = x >> 3;
     c = bloom->bf[byte];        // expensive memory access
